refined code for checking authentication for valid user in their posts

+++ src/controller/jobController.js
@@ -72,25 +72,46 @@ exports.createJob = async (req, res, next) => {
 
 exports.deleteJob = async (req, res, next) => {
   try {
-    const {
-      id,
-    } = req.params;
-    const findJob = await Job.findOne({
-      _id: id,
-    });
-    if (findJob) {
-      const deleteJob = await Job.findOneAndDelete({
+    if (req.userData.userRole === 'employers') {
+      const {
+        id,
+      } = req.params;
+      const findJob = await Job.findOne({
         _id: id,
       });
-      if (deleteJob) {
-        res.status(200).json({
-          message: 'Job deleted successfully',
+      if (findJob) {
+        const employerId = req.userData.id;
+        const checkJobCreatedUserId = await Job.findOne(
+          {
+            $and: [
+              { _id: id }, { userId: employerId },
+            ],
+          },
+        );
+        if (checkJobCreatedUserId) {
+          const deleteJob = await Job.findOneAndDelete({
+            _id: id,
+          });
+          if (deleteJob) {
+            res.status(200).json({
+              message: 'Job deleted successfully',
+            });
+          }
+        }
+        if (!checkJobCreatedUserId) {
+          res.status(401).json({
+            message: 'unauthorized user',
+          });
+        }
+      }
+      if (!findJob) {
+        res.status(400).json({
+          message: 'Job do not exists',
         });
       }
-    }
-    if (!findJob) {
-      res.status(400).json({
-        message: 'Job do not exists',
+    } else {
+      res.status(401).json({
+        message: 'you are not a employer!',
       });
     }
   } catch (error) {
@@ -100,60 +121,88 @@ exports.deleteJob = async (req, res, next) => {
 
 exports.editJob = async (req, res, next) => {
   try {
-    const validationSchema = joi.object({
-      jobName: joi.string().trim().required(),
-      jobSalary: joi.string().trim().required(),
-      jobCategory: joi.string().trim().required(),
-      jobCompanyName: joi.string().trim().required(),
-      jobDescription: joi.string().trim().min(10).max(500)
-        .required(),
-      jobType: joi.string().trim().required(),
-      jobRequiredSkills: joi.array().required(),
-    });
-    const validate = await validationSchema.validateAsync(req.body);
-    if (validate) {
-      const {
-        jobName,
-        jobSalary,
-        jobCategory,
-        jobCompanyName,
-        jobDescription,
-        jobType,
-        jobRequiredSkills,
-      } = req.body;
+    if (req.userData.userRole === 'employers') {
       const {
         id,
       } = req.params;
-      const checkJob = await Job.findOne({
+      const findJob = await Job.findOne({
         _id: id,
       });
-      const editJobInfo = await Job.updateOne({
-        _id: id,
-      }, {
-        $set: {
-          jobName,
-          jobSalary,
-          jobCategory,
-          jobCompanyName,
-          jobDescription,
-          jobType,
-          jobRequiredSkills,
-        },
-      });
-      if (editJobInfo.nModified === 0) {
-        res.status(304).json({
-          message: 'Job information not modified',
-        });
-      } else {
-        res.status(201).json({
-          message: 'Job information edited successfully',
-        });
+      if (findJob) {
+        const employerId = req.userData.id;
+        const checkJobCreatedUserId = await Job.findOne(
+          {
+            $and: [
+              { _id: id }, { userId: employerId },
+            ],
+          },
+        );
+        if (checkJobCreatedUserId) {
+          const validationSchema = joi.object({
+            jobName: joi.string().trim().required(),
+            jobSalary: joi.string().trim().required(),
+            jobCategory: joi.string().trim().required(),
+            jobCompanyName: joi.string().trim().required(),
+            jobDescription: joi.string().trim().min(10).max(500)
+              .required(),
+            jobType: joi.string().trim().required(),
+            jobRequiredSkills: joi.array().required(),
+          });
+          const validate = await validationSchema.validateAsync(req.body);
+          if (validate) {
+            const {
+              jobName,
+              jobSalary,
+              jobCategory,
+              jobCompanyName,
+              jobDescription,
+              jobType,
+              jobRequiredSkills,
+            } = req.body;
+            const editJobInfo = await Job.updateOne({
+              _id: id,
+            }, {
+              $set: {
+                jobName,
+                jobSalary,
+                jobCategory,
+                jobCompanyName,
+                jobDescription,
+                jobType,
+                jobRequiredSkills,
+              },
+            });
+            if (editJobInfo.nModified === 0) {
+              res.status(304).json({
+                message: 'Job information not modified',
+              });
+            } else {
+              res.status(201).json({
+                message: 'Job information edited successfully',
+              });
+            }
+          }
+          if (!validate) {
+            res.status(422).json({
+              message: 'validation error, check your data entries!',
+            });
+          }
+        }
+        if (!checkJobCreatedUserId) {
+          res.status(401).json({
+            message: 'unauthorized user',
+          });
+        }
       }
-      if (!checkJob) {
+      if (!findJob) {
         res.status(400).json({
-          message: 'Job do not exist',
+          message: 'Job do not exists',
         });
       }
+    } else {
+      res.status(401).json({
+        message: 'you are not a employer!',
+      });
     }
   } catch (error) {
     if (error.name === 'ValdationError') {
